+++ changes.md
@@ -8,6 +8,8 @@
 
 ### Bug fixes
 
+* Interactive decorations (e.g., Archaeologist's Office) now appear in Special Buildings instead of Decorations in the scanner
+
 ### Internal
 
 ## v1.2.3